+++ src/components/Nav.js
@@ -14,21 +14,6 @@ const Nav = () => {
     return (
         <nav className="nav">
             <ul className="nav__list">
-                {/*<li className="nav__item">*/}
-                {/*    <Link*/}
-                {/*        to="/"*/}
-                {/*        className="nav__link"*/}
-                {/*        onClick={changeBasketModalStatus}*/}
-                {/*    >*/}
-                {/*        Піца*/}
-                {/*    </Link>*/}
-                {/*</li>*/}
-                {/*<li className="nav__item">*/}
-                {/*    <Link to="drink-page" className="nav__link">Напої</Link>*/}
-                {/*</li>*/}
-                {/*<li className="nav__item">*/}
-                {/*    <Link to="salat-page" className="nav__link">Салати</Link>*/}
-                {/*</li>*/}
                 {
                     tabs.map(tab => (
                         <li className="nav__item" key={tab.name}>
