+++ src/components/UI/index.js
@@ -1,3 +1,13 @@
-import ButtonComponet from './button/ButtonComponet';
-import ModalComponent from './modal/ModalComponent';
-import ModalBasket from './modal/ModalBasket';
+export ButtonComponet from "./button/ButtonComponet";
+export Card from "./cards/Card";
+export DiscountCard from "./cards/DiscountCard";
+export DrinkCard from "./cards/DrinkCard";
+
+export SliderAlso from "./carousel/SliderAlso";
+
+export InputComponet from "./input/inputComponent";
+
+export ModalComponent from "./modal/ModalComponent";
+export ModalBasket from "./modal/ModalBasket";
+export BasketForm from "./modal/BasketForm";
+export SmallCard from "./modal/SmallCard";
