+++ src/components/UI/modal/BasketForm.js
@@ -1,12 +1,12 @@
 import React from "react";
 import {useSelector} from "react-redux";
-import {useActions} from "../../../helpers/hooks/useActions";
-import "./modal.scss";
-import SmallCard from "./SmallCard";
 import {useEffect} from "react";
+import {useActions} from "../../../helpers/hooks/useActions";
 import {fetchProductToBasket} from "../../../redux/actions/basketAction";
-const BasketForm = () => {
+import SmallCard from "./SmallCard";
+import "./modal.scss";
 
+const BasketForm = () => {
     const basketDataItems = useSelector(state => state.basket.basketData);
 
     const getProducts = useActions(fetchProductToBasket)
