+++ src/components/UI/modal/SmallCard.js
@@ -1,9 +1,8 @@
 import React from "react";
 import {Link} from "react-router-dom";
-
-import "./modal.scss";
 import {useActions} from "../../../helpers/hooks/useActions";
 import { deleteProductFromBasket} from "../../../redux/actions/basketAction";
+import "./modal.scss";
 
 const SmallCard = (props) => {
     const {type, id, img, title, description, size, weight, price} = props;
@@ -11,10 +10,10 @@ const SmallCard = (props) => {
 
     const removeProductFromBasket = useActions(deleteProductFromBasket);
 
-
     const deleteFromBasket = () => {
         removeProductFromBasket(props);
     }
+
     return (
         <div className="basket-item">
             <Link to="/">
@@ -40,4 +39,3 @@ const SmallCard = (props) => {
 }
 
 export default SmallCard;
-
