+++ src/index.js
@@ -6,7 +6,6 @@ import App from './App';
 import store from "./store";
 import { Provider } from "react-redux";
 
-
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <Provider store={store}>
@@ -15,4 +14,3 @@ root.render(
       </BrowserRouter>
   </Provider>
 );
-
