+++ src/redux/actions/basketAction.js
@@ -111,4 +111,3 @@ export function* basketWatcher(){
         fork(closeBasketModalStatusWatcher),
     ])
 }
-
