+++ src/redux/actions/salatActions.js
@@ -2,7 +2,6 @@ import { all, fork, takeEvery, call, put } from "redux-saga/effects";
 import { FETCH_SALAT, FETCH_SALAT_SUCCESS } from "../types";
 import { ApiService } from "../../helpers/api-service";
 
-
 export const fetchSalat = () => ({
     type: FETCH_SALAT,
 });
